8: Separate in App into LoginPage and LoggedInArea
