+++ src/frontend/LoginPage/LoginPage.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import { Routes } from '../shared/config';
+import { RoutePaths } from '../shared/config';
 
 export default class LoginPage extends React.Component {
     render() {
         return (
-            <Link to={ Routes.ChessGamePage }>Play!</Link>
+            <Link to={ RoutePaths.ChessGamePage }>Play!</Link>
         )
     }
 }
