+++ src/frontend/shared/config.ts
@@ -1,4 +1,5 @@
-export enum Routes {
+export enum RoutePaths {
     LoginPage = '/',
+    LoggedInArea = '/(.+)',
     ChessGamePage = '/play'
 }
